Move Door related staff to separate folder
