+++ Elements.Serialization.IFC/src/Serialization/IFC/IFCExtensions.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Elements;
+using Elements.BuildingElements.Door;
 using Elements.Geometry;
 using Elements.Geometry.Interfaces;
 using Elements.Geometry.Solids;
