+++ Elements.Serialization.IFC/test/IFCTests.cs
@@ -8,6 +8,7 @@ using Xunit.Abstractions;
 using System.Collections.Generic;
 using Elements.Geometry.Profiles;
 using System.Linq;
+using Elements.BuildingElements.Door;
 
 namespace Elements.IFC.Tests
 {
