+++ Elements/test/DoorTest.cs
@@ -1,4 +1,5 @@
-﻿using Elements.Geometry;
+﻿using Elements.BuildingElements.Door;
+using Elements.Geometry;
 using Elements.Tests;
 using System;
 using System.Collections.Generic;
